Move validate_alias function to validators file
