+++ python/nav/django/validators.py
@@ -22,6 +22,8 @@ from decimal import Decimal, InvalidOperation
 from django.utils.translation import gettext
 from django.core.exceptions import ValidationError
 
+MAX_ALIAS_LENGTH = 64
+
 
 def is_valid_point_string(point_string):
     if len(point_string.split(',')) == 2:
@@ -74,3 +76,28 @@ def validate_hstore(value):
     dictionary = json.loads(value)
 
     return dictionary
+
+
+def validate_aliases(aliases: list[str]) -> list[str]:
+    """
+    Validates a given list of aliases and raises a ValidationError if any of the
+    aliases contain the pipe character or are too long
+
+    Returns a deduplicated and stripped version of the given list
+    """
+    if not aliases:
+        return []
+    cleaned = []
+    for item in aliases:
+        if not isinstance(item, str):
+            raise ValidationError("All aliases must be strings.")
+        stripped = item.strip()
+        if "|" in stripped:
+            raise ValidationError("Aliases cannot contain the pipe character ('|')")
+        if len(stripped) > MAX_ALIAS_LENGTH:
+            raise ValidationError(
+                f"Alias must be {MAX_ALIAS_LENGTH} characters or fewer."
+            )
+        if stripped and stripped not in cleaned:
+            cleaned.append(stripped)
+    return cleaned
